feat(projets): replace project overlay with dedicated detail page
Closes #106. Cards on /projets and the home "à la une" section now link
to /projets/<id>, a full-page project detail (title, tags, year/duration,
credits, description, YouTube embed) with a back link and a 404 state
for unknown ids. ProjectOverlay is removed; useOverlay and the shared
content-overlay__* styles stay since ArticleOverlay still uses them.
