+++ frontend/app/globals.css
@@ -403,6 +403,7 @@ body {
   overflow: hidden;
   transition: transform 0.3s ease;
   color: #fff;
+  text-decoration: none;
   font-family: inherit;
   animation: filterCardIn 0.35s ease both;
 }
@@ -483,14 +484,33 @@ body {
   background: rgba(255, 255, 255, 0.14);
 }
 
-/* Project overlay video */
-.project-overlay__video {
+/* Project detail page */
+.project-detail {
+  max-width: 800px;
+  margin: 1rem 0 0;
+}
+
+.project-detail__back {
+  display: inline-block;
+  margin-bottom: 1.5rem;
+  color: #fff;
+  text-decoration: none;
+  font-size: 0.95rem;
+  opacity: 0.7;
+  transition: opacity 0.2s ease;
+}
+
+.project-detail__back:hover {
+  opacity: 1;
+}
+
+.project-detail__video {
   width: 100%;
   aspect-ratio: 16 / 9;
   margin-top: 2rem;
 }
 
-.project-overlay__video iframe {
+.project-detail__video iframe {
   width: 100%;
   height: 100%;
   border: none;
